separate Container/Layout component
